refactor: PaymentCommandService를 통해 결제 승인 API 연동 수정
